+++ packages/shared/src/index.ts
@@ -1,14 +1,20 @@
 export { FlowgladActionKey, HTTPMethod } from './types'
 export {
   createCheckoutSessionSchema,
+  createProductCheckoutSessionSchema,
+  createAddPaymentMethodCheckoutSessionSchema,
   cancelSubscriptionSchema,
   flowgladActionValidators,
   createUsageEventSchema,
+  createSubscriptionSchema,
 } from './actions'
 export type {
   FlowgladActionValidatorMap,
   CreateCheckoutSessionParams,
+  CreateProductCheckoutSessionParams,
+  CreateAddPaymentMethodCheckoutSessionParams,
   CancelSubscriptionParams,
   CreateUsageEventParams,
+  CreateSubscriptionParams,
 } from './actions'
 export { getBaseURL } from './utils'